rename(christmas): Controller의 input 디렉토리 삭제
